+++ src/Security/UserChecker.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Security;
+
+use App\Entity\User;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Security\Core\User\UserCheckerInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
+
+final readonly class UserChecker implements UserCheckerInterface
+{
+    public function checkPreAuth(UserInterface $user): void
+    {
+        if (!$user instanceof User) {
+            return;
+        }
+        if (!$user->isEnabled()) {
+            throw new AccessDeniedException('Your user account is not enabled.');
+        }
+    }
+
+    public function checkPostAuth(UserInterface $user): void
+    {
+        if (!$user instanceof User) {
+            return;
+        }
+        if (!$user->isEnabled()) {
+            throw new AccessDeniedException('Your user account is not enabled.');
+        }
+    }
+}