+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ##### Version 5.8.03 (WIP)
  * fix bugs
  * add more ROLE user cases into testing suite
+ * add UserChecker to prevent not enabled User logins
 
 ##### Version 5.8.02 (2025-03-01)
  * fix missing composer require symfony/doctrine-messenger dependency problem